. Add a basic tsconfig.json to the slt-bable-plugin.
This is to accomindate the typescript compiler, which we were receiving
an error that this file was missing.

Adding this file resolves that error, however we have a new error, which
is that the compiler cant find any inputs

    slt-js-monorepo 🍀 pnpm -w build
    > slt-js@1.0.0 build /data/data/com.termux/files/home/slt-js-monorepo
    > pnpm -r --parallel build

    Scope: 6 of 7 workspace projects
    packages/slt-js-core build$ tsc -b
    packages/slt-babel-plugin build$ tsc -p tsconfig.json || echo build not configured
    packages/slt-acorn-plugin build$ tsc -b
    packages/slt-babel-plugin build: error TS5083: Cannot read file '/data/data/com.termux/files/home/slt-js-monorepo/tsconfig.base.json'.
    packages/slt-babel-plugin build: error TS18003: No inputs were found in config file '/data/data/com.termux/files/home/slt-js-monorepo/packages/slt-babel-plugin/tsconfig.json'. Specified 'include' paths were '["src"]' and 'exclude' paths were '["/data/data/com.termux/files/home/slt-js-monorepo/packages/slt-babel-plugin/dist"]'.
    packages/slt-babel-plugin build: build not configured
    packages/slt-js-core build: Done
    packages/slt-babel-plugin build: Done
    packages/slt-acorn-plugin build: Done

This is due in part to tgere not being an src directory. there is a lib directory, which we have not referenced in this version of the tscondif, but even if we had, we get same error, because that doesn't include any ts files either.

Perhap js is forward compatiable with ts, and we can
just rename the extention to ts to satisfy the
compiler, but that seems like putting the ass before
the donkey, because ts is just going to compile back
to js, so why bother if we already have js, unless we
really really need the type system for some reason.
In this case we can probably side step it entirely,
as this was probably an oversight, but i don't know
just documenting what i'm seeing for a basis for
improvement on this portion of the project.. 💗