. Remove build task from babl plugin.
At least for now. It doesn't seem to be need for the code which is
present on this head, and gets rid of the error which shows up when
building the full worspace with pnpm -w build. It's probably a good idea
to cremove the tsconfig.json and any other ts artifacts we have in there
if we aren't typescript in the babel, or have no need for it there, to
avoid confusion, but have done that in this commit, just satisfying the
workspace build. Look Mom, no errors! 😁🎀

    slt-js-monorepo 🍀 pnpm -w build                                                                            > slt-js@1.0.0 build /data/data/com.termux/files/home/slt-js-monorepo
    > pnpm -r --parallel build                                                                                  Scope: 6 of 7 workspace projects
    packages/slt-acorn-plugin build$ tsc -b               packages/slt-js-core build$ tsc -b
    packages/slt-acorn-plugin build: Done
    packages/slt-js-core build: Done